AK: Fix overflow and mixed-signedness issues in binary_search() (#2961)

+++ Kernel/VM/RangeAllocator.cpp
@@ -172,7 +172,7 @@ void RangeAllocator::deallocate(Range range)
 
     ASSERT(!m_available_ranges.is_empty());
 
-    int nearby_index = 0;
+    size_t nearby_index = 0;
     auto* existing_range = binary_search(
         m_available_ranges.span(), range, [](auto& a, auto& b) {
             return a.base().get() - b.end().get();
